azure_flask_webapp/requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
pandas
sqlalchemy
psycopg2-binary
python-dotenv
plotly
